requirements.txt
s3fs
prefect_slack
prefect_shell
sqlalchemy
pandas
awswrangler